moved test to its own file test
