+++ tests/Service/AtozTitlesService/FindAvailableTleosByFirstLetterTest.php
@@ -45,17 +45,4 @@ class FindAvailableTleosByFirstLetterTest extends AbstractAtozTitlesServiceTest
 
         $this->assertEquals([], $atozTitles);
     }
-
-
-    public function testCountAvailableTleosByFirstLetter()
-    {
-        $this->mockRepository->expects($this->once())
-            ->method('countTleosByFirstLetter')
-            ->with('t')
-            ->willReturn(10);
-
-        $result = $this->service()->countAvailableTleosByFirstLetter('t');
-
-        $this->assertEquals(10, $result);
-    }
 }
